+++ app.py
@@ -1,45 +1,6 @@
-from flask import Flask
-from flask_login import LoginManager
-import os
-from models import db, User
-from auth import auth
-from profile import profile
-from library import library
-from stories import stories
-from search import search
+from app import create_app
 
-def create_app():
-    app = Flask(__name__)
-    app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY', 'dev')
-    app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URL', 'postgresql://postgres:postgres@db:5432/scribe')
-    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-    app.config['UPLOAD_FOLDER'] = 'static/uploads'
-    app.config['MAX_CONTENT_LENGTH'] = 16 * 1024 * 1024  # 16MB max file size
-    app.config['COVER_SIZE'] = (512, 800)  # Width, Height
-
-    # Initialize extensions
-    db.init_app(app)
-    login_manager = LoginManager()
-    login_manager.init_app(app)
-    login_manager.login_view = 'auth.login'
-
-    @login_manager.user_loader
-    def load_user(user_id):
-        return User.query.get(int(user_id))
-
-    # Register blueprints
-    app.register_blueprint(auth)
-    app.register_blueprint(profile)
-    app.register_blueprint(library)
-    app.register_blueprint(stories)
-    app.register_blueprint(search)
-
-    # Create database tables
-    with app.app_context():
-        db.create_all()
-
-    return app
+app = create_app()
 
 if __name__ == '__main__':
-    app = create_app()
     app.run(host='0.0.0.0', debug=True)
