+++ app/auth.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, render_template, request, redirect, url_for, flash
 from flask_login import login_user, logout_user, login_required
 from werkzeug.security import generate_password_hash, check_password_hash
-from models import db, User
+from app.models import db, User
 
 auth = Blueprint('auth', __name__)
