+++ app/library.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, render_template, redirect, url_for, flash
 from flask_login import login_required, current_user
-from models import db, SavedStory, Story
+from app.models import db, SavedStory, Story
 
 library = Blueprint('library', __name__)
