+++ app/profile.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, render_template, request, redirect, url_for, flash
 from flask_login import login_required, current_user
-from models import db, User, Story, Rating
+from app.models import db, User, Story, Rating
 
 profile = Blueprint('profile', __name__)
